Fix incorrect rounding of gps coordinates in Degrees/Minutes format

+++ src/core/numericformats/qgscoordinatenumericformat.cpp
@@ -200,7 +200,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLatitudeAsDegreesMinutesSeco
     wrappedY = wrappedY + 180.0;
   }
 
-  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   int degreesY = int( std::fabs( wrappedY ) );
   const double floatMinutesY = ( std::fabs( wrappedY ) - degreesY ) * 60.0;
@@ -291,7 +291,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLongitudeAsDegreesMinutesSec
     wrappedX = wrappedX + 360.0;
   }
 
-  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   int degreesX = int( std::fabs( wrappedX ) );
   const double floatMinutesX = ( std::fabs( wrappedX ) - degreesX ) * 60.0;
@@ -391,7 +391,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLatitudeAsDegreesMinutes( do
   int degreesY = int( std::fabs( wrappedY ) );
   double floatMinutesY = ( std::fabs( wrappedY ) - degreesY ) * 60.0;
 
-  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   //make sure rounding to specified precision doesn't create minutes >= 60
   if ( std::round( floatMinutesY * precisionMultiplier ) >= 60 * precisionMultiplier )
@@ -462,7 +462,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLongitudeAsDegreesMinutes( d
   int degreesX = int( std::fabs( wrappedX ) );
   double floatMinutesX = ( std::fabs( wrappedX ) - degreesX ) * 60.0;
 
-  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   //make sure rounding to specified precision doesn't create minutes >= 60
   if ( std::round( floatMinutesX * precisionMultiplier ) >= 60 * precisionMultiplier )
@@ -538,7 +538,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLatitudeAsDegrees( double va
 
   const double absY = std::fabs( wrappedY );
 
-  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   QString hemisphere;
   QString sign;
@@ -590,7 +590,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLongitudeAsDegrees( double v
 
   const double absX = std::fabs( wrappedX );
 
-  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   QString hemisphere;
   QString sign;
